+++ lib/punch/services.rb
@@ -1,4 +1,5 @@
 require_relative "services/punch_sentry"
 require_relative "services/punch_source"
+require_relative "services/punch_plugin"
 require_relative "services/punch_domain"
 require_relative "services/report_status"