+++ test/punch/basics/test_plugin.rb
@@ -2,23 +2,23 @@ require_relative "../../test_helper"
 
 describe Plugin do
   class Dummy
-    extend Plugin
+    extend Punch::Plugin
   end
 
   let(:holder) {
     Module.new do
-      extend Plugin::Holder
+      extend Punch::Plugin::Holder
       plugin Dummy
     end
   }
 
   it '#plugin must return Pluging::Holder' do
-    assert_kind_of Plugin::Holder, Dummy.plugin
+    assert_kind_of Punch::Plugin::Holder, Dummy.plugin
   end
 
   describe Plugin::Holder do
     it 'must be Plugin::Holder' do
-      assert_kind_of Plugin::Holder, holder
+      assert_kind_of Punch::Plugin::Holder, holder
       assert_respond_to holder, :plugin
       assert_respond_to holder, :object
       assert_respond_to holder, :object=