+++ test/punch/services/test_punch_domain.rb
@@ -10,5 +10,25 @@ describe PunchDomain do
       domain = build_sample_domain
       service.(domain)
     }
+
+    Sandbox.() {
+      @config = Punch::Config.new('lib', 'test', 'dom', 'sen', 'ent', 'ser', 'plg')
+      @domain = build_sample_domain
+      Punch.stub :config, @config do
+        service.(@domain)
+      end
+      # payload = [
+      #   "lib/dom/ser/admin_lock_user.rb",
+      #   "test/dom/ser/test_admin_lock_user.rb",
+      #   "lib/dom/plg.rb",
+      #   "lib/dom/plg/secrets.rb",
+      #   "test/dom/plg/test_secrets.rb",
+      #   "lib/dom/sen.rb",
+      #   "lib/dom.rb",
+      # ]
+      # puts "-= Custom Configuration =-"
+      # print_folders
+      # print_content(*payload)
+    }
   end
 end